restructure post_objects
